Move classes into relevant packages
