+++ src/main/java/bob/Bob.java
@@ -1,11 +1,15 @@
 package bob;
 
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.List;
+
+import bob.command.Command;
+import bob.exception.EmptyFieldException;
+import bob.exception.InvalidCommandException;
+import bob.task.Task;
 
 
 public class Bob {
