+++ src/main/java/bob/Parser.java
@@ -3,6 +3,22 @@ package bob;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
+import bob.command.AddCommand;
+import bob.command.Command;
+import bob.command.DeleteCommand;
+import bob.command.FindCommand;
+import bob.command.HelpCommand;
+import bob.command.ListCommand;
+import bob.command.MarkCommand;
+import bob.command.MatchListCommand;
+import bob.command.UnmarkCommand;
+import bob.exception.EmptyFieldException;
+import bob.exception.InvalidCommandException;
+import bob.task.Deadline;
+import bob.task.Event;
+import bob.task.Task;
+import bob.task.ToDo;
+
 /**
  * Represents Bob's parser.
  */
