+++ src/main/java/bob/command/Command.java
@@ -1,4 +1,7 @@
-package bob;
+package bob.command;
+
+import bob.Storage;
+import bob.TaskList;
 
 public abstract class Command {
     public abstract String execute(TaskList tasks, Storage storage);
