+++ src/main/java/bob/command/DeleteCommand.java
@@ -1,8 +1,12 @@
-package bob;
+package bob.command;
 
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+import bob.Storage;
+import bob.TaskList;
+import bob.task.Task;
+
+public class DeleteCommand extends Command {
     private int taskIndex;
 
     /**
@@ -10,19 +14,22 @@ public class UnmarkCommand extends Command {
      *
      * @param taskIndex index of task.
      */
-    public UnmarkCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     @Override
     public String execute(TaskList tasks, Storage storage) {
         try {
-            Task task = tasks.getTask(taskIndex);
-            task.unmark();
+            Task task = tasks.remove(taskIndex);
             storage.save(tasks);
-            return String.format("Oh well, this task has been marked undone:\n\t%s", task);
+            return String.format("""
+                    Oof. I have removed the requested task:
+                    \t%s
+                    Now you have %s tasks in the list""", task, tasks.size());
+
         } catch (IndexOutOfBoundsException e) {
-            return "There's no such task!";
+            return "Hm, you don't seem to have that task";
         } catch (IOException e) {
             return "I can't remember that for some reason T T";
         }
