+++ src/main/java/bob/command/MatchListCommand.java
@@ -1,4 +1,7 @@
-package bob;
+package bob.command;
+
+import bob.Storage;
+import bob.TaskList;
 
 public class MatchListCommand extends Command {
     @Override
