+++ src/main/java/bob/exception/EmptyFieldException.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.exception;
 
 public class EmptyFieldException extends BobException {
 }
