+++ src/main/java/bob/exception/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.exception;
 
 /**
  * Thrown when command is invalid.
