+++ src/main/java/bob/task/Event.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
