+++ src/main/java/bob/task/Task.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.task;
 
 public abstract class Task {
     protected String description;
