+++ src/main/java/bob/task/ToDo.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
